chore(django-users): create an App to users & JWT (#22)
* chore(django-users): create an App to users & CONTRIBUTING guide

* Include only necessary urls from allauth

* Remove unnecessary templates

* Add X-UID-Key, custom authentication for API calls

* Add users.auth module and disable csfr authentication on some API calls

* Add 404 error handler and start working on the User page

* Display user UUID in profile when logged with user

* About, Docs, API Docs pages on navbar. Hide API calls from Swagger. Disable posting directly to db from Swagger

* Create Author at User's creation

* Inherit base.html on swagger page (/api/docs/)

* Update makim dev.dotenv

* Fix TTY checker on makin dev.dotenv

* Add Github Oauth provisory secrets on CI

* Try inheriting the github secrets

+++ src/registry/api.py
@@ -1,12 +1,14 @@
 import datetime
 from typing import List
 
+from django.contrib.auth import get_user_model
+from django.views.decorators.csrf import csrf_exempt
 from ninja import Query, Router
 from ninja.orm.fields import AnyObject
 from ninja.security import django_auth
 
-from main.models import CustomUser
 from main.schema import ForbiddenSchema, NotFoundSchema, Schema, SuccessSchema
+from users.auth import UidKeyAuth
 
 from .models import Author, Model, Prediction
 from .schema import (
@@ -14,11 +16,14 @@ from .schema import (
     AuthorSchema,
     ModelFilterSchema,
     ModelSchema,
-    PredictionSchema,
     PredictionFilterSchema,
+    PredictionSchema,
 )
+from .utils import calling_via_swagger
 
 router = Router()
+uidkey = UidKeyAuth()
+User = get_user_model()
 
 
 # [Model] Author
@@ -35,7 +40,8 @@ class AuthorInPost(Schema):
     institution: str
 
 
-@router.get("/authors/", response=List[AuthorSchema])
+@router.get("/authors/", response=List[AuthorSchema], tags=["registry", "authors"])
+@csrf_exempt
 def list_authors(
     request,
     filters: AuthorFilterSchema = Query(...),
@@ -46,7 +52,12 @@ def list_authors(
     return authors
 
 
-@router.get("/authors/{username}", response={200: AuthorSchema, 404: NotFoundSchema})
+@router.get(
+    "/authors/{username}",
+    response={200: AuthorSchema, 404: NotFoundSchema},
+    tags=["registry", "authors"],
+)
+@csrf_exempt
 def get_author(request, username: str):
     """Gets author by Github username"""
     try:
@@ -60,11 +71,16 @@ def get_author(request, username: str):
     "/authors/",
     response={201: AuthorSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
     auth=django_auth,
+    tags=["registry", "authors"],
+    include_in_schema=False,
 )
 def create_author(request, payload: AuthorIn):
-    """Posts author to database, requires a CustomUser to be created"""
+    """
+    Posts author to database, requires a User to be created
+    @note: This call is related to User and shouldn't be done only via API Call
+    """
     try:
-        user = CustomUser.objects.get(username=payload.user)
+        user = User.objects.get(username=payload.user)
         try:
             author = Author.objects.get(user__username=payload.user)
             if author:
@@ -72,7 +88,7 @@ def create_author(request, payload: AuthorIn):
         except Author.DoesNotExist:
             author = Author.objects.create(user=user, institution=payload.institution)
             return 201, author
-    except CustomUser.DoesNotExist:
+    except User.DoesNotExist:
         return 404, {"message": f"User '{payload.user}' does not exist"}
 
 
@@ -80,6 +96,8 @@ def create_author(request, payload: AuthorIn):
     "/authors/{username}",
     response={201: AuthorSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
     auth=django_auth,
+    tags=["registry", "authors"],
+    include_in_schema=False,
 )
 def update_author(request, username: str, payload: AuthorInPost):
     """
@@ -93,7 +111,11 @@ def update_author(request, username: str, payload: AuthorInPost):
             return 403, {"message": "You are not authorized to update this author"}
 
         author.institution = payload.institution
-        author.save()
+
+        if not calling_via_swagger(request):
+            # Not realy required, since include_in_schema=False
+            author.save()
+
         return 201, author
     except Author.DoesNotExist:
         return 404, {"message": "Author not found"}
@@ -103,16 +125,24 @@ def update_author(request, username: str, payload: AuthorInPost):
     "/authors/{username}",
     response={200: SuccessSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
     auth=django_auth,
+    tags=["registry", "authors"],
+    include_in_schema=False,
 )
 def delete_author(request, username: str):
-    """Deletes author"""
+    """
+    Deletes author
+    @note: This call is related to User and shouldn't be done only via API Call
+    """
     try:
         author = Author.objects.get(user__username=username)
 
         if request.user != author.user:  # TODO: Enable admins here
             return 403, {"message": "You are not authorized to delete this author"}
 
-        author.delete()
+        if not calling_via_swagger(request):
+            # Not realy required, since include_in_schema=False
+            author.delete()
+
         return 200, {"message": f"Author '{author.user.name}' deleted successfully"}
     except Author.DoesNotExist:
         return 404, {"message": "Author not found"}
@@ -122,20 +152,34 @@ def delete_author(request, username: str):
 class ModelIn(Schema):
     name: str
     description: str = None
-    author: str  # Author username
     repository: str  # TODO: Validate repository?
     implementation_language: str
     type: str
 
 
-@router.get("/models/", response=List[ModelSchema])
+class ModelInUpdate(ModelIn):
+    """
+    Enable User to update Model's author.
+    With this operation, the user will lose model's permissions
+    """
+
+    author: str
+
+
+@router.get("/models/", response=List[ModelSchema], tags=["registry", "models"])
+@csrf_exempt
 def list_models(request, filters: ModelFilterSchema = Query(...)):
     models = Model.objects.all()
     models = filters.filter(models)
     return models
 
 
-@router.get("/models/{model_id}", response={200: ModelSchema, 404: NotFoundSchema})
+@router.get(
+    "/models/{model_id}",
+    response={200: ModelSchema, 404: NotFoundSchema},
+    tags=["registry", "models"],
+)
+@csrf_exempt
 def get_model(request, model_id: int):
     try:
         model = Model.objects.get(pk=model_id)  # TODO: get model by id?
@@ -146,20 +190,17 @@ def get_model(request, model_id: int):
 
 @router.post(
     "/models/",
-    response={201: ModelSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
-    auth=django_auth,
+    response={201: ModelSchema},
+    auth=uidkey,
+    tags=["registry", "models"],
 )
+@csrf_exempt
 def create_model(request, payload: ModelIn):
-    try:
-        author = Author.objects.get(user__username=payload.author)
-        if request.user != author.user:
-            return 403, {
-                "message": "You are not authorized to add a Model to this author"
-            }
-        payload.author = author
-    except Author.DoesNotExist:
-        return 404, {"message": "Invalid Author"}
-    model = Model.objects.create(**payload.dict())
+    uid, _ = request.headers.get("X-UID-Key").split(":")
+    author = Author.objects.get(user__username=uid)
+    model = Model(author=author, **payload.dict())
+    if not calling_via_swagger(request):
+        model.save()
     return 201, model
 
 
@@ -167,8 +208,10 @@ def create_model(request, payload: ModelIn):
     "/models/{model_id}",
     response={201: ModelSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
     auth=django_auth,
+    tags=["registry", "models"],
+    include_in_schema=False,
 )
-def update_model(request, model_id: int, payload: ModelIn):
+def update_model(request, model_id: int, payload: ModelInUpdate):
     try:
         model = Model.objects.get(pk=model_id)  # TODO: Update by id?
 
@@ -176,13 +219,16 @@ def update_model(request, model_id: int, payload: ModelIn):
             return 403, {"message": "You are not authorized to update this Model"}
 
         try:
-            author = Author.objects.get(user__username=payload.author)
+            author = Author.objects.get(user__username=request.user)
             payload.author = author
 
             for attr, value in payload.dict().items():
                 setattr(model, attr, value)
 
-            model.save()
+            if not calling_via_swagger(request):
+                # Not realy required, since include_in_schema=False
+                model.save()
+
             return 201, model
         except Author.DoesNotExist:
             return 404, {
@@ -198,6 +244,8 @@ def update_model(request, model_id: int, payload: ModelIn):
     "/models/{model_id}",
     response={204: SuccessSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
     auth=django_auth,
+    tags=["registry", "models"],
+    include_in_schema=False,
 )
 def delete_model(request, model_id: int):
     try:
@@ -206,7 +254,10 @@ def delete_model(request, model_id: int):
         if request.user != model.author.user:
             return 403, {"message": "You are not authorized to delete this Model"}
 
-        model.delete()
+        if not calling_via_swagger(request):
+            # Not realy required, since include_in_schema=False
+            model.delete()
+
         return 204, {"message": f"Model {model.name} deleted successfully"}
     except Author.DoesNotExist:
         return 404, {"message": "Model not found"}
@@ -214,14 +265,17 @@ def delete_model(request, model_id: int):
 
 # [Model] Prediction
 class PredictionIn(Schema):
-    model: ModelSchema  # TODO: change it. Issue #20
+    model: int  # TODO: change it. Issue #20
     description: str = None
     commit: str
-    predict_date: datetime.date
+    predict_date: datetime.date  # YYYY-mm-dd
     prediction: AnyObject
 
 
-@router.get("/predictions/", response=List[PredictionSchema])
+@router.get(
+    "/predictions/", response=List[PredictionSchema], tags=["registry", "predictions"]
+)
+@csrf_exempt
 def list_predictions(
     request,
     filters: PredictionFilterSchema = Query(...),
@@ -234,7 +288,9 @@ def list_predictions(
 @router.get(
     "/predictions/{predict_id}",
     response={200: PredictionSchema, 404: NotFoundSchema},
+    tags=["registry", "predictions"],
 )
+@csrf_exempt
 def get_prediction(request, predict_id: int):
     try:
         prediction = Prediction.objects.get(pk=predict_id)  # TODO: get by id?
@@ -246,29 +302,32 @@ def get_prediction(request, predict_id: int):
 @router.post(
     "/predictions/",
     response={201: PredictionSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
-    auth=django_auth,
+    auth=uidkey,
+    tags=["registry", "predictions"],
 )
+@csrf_exempt
 def create_prediction(request, payload: PredictionIn):
     try:
-        model = Model.objects.get(name=payload.model)  # TODO: get by name?
-
-        if request.user != model.author.user:
-            return 403, {
-                "message": "You are not authorized to add a prediction to this model"
-            }
-
-        payload.model = model
-        # TODO: Add commit verification here #19
-        prediction = Prediction.objects.create(**payload.dict())
-        return 201, prediction
+        model = Model.objects.get(pk=payload.model)  # TODO: get by id?
     except Model.DoesNotExist:
         return 404, {"message": f"Model '{payload.model}' not found"}
+
+    payload.model = model
+    # TODO: Add commit verification here #19
+    prediction = Prediction(**payload.dict())
+
+    if not calling_via_swagger(request):
+        prediction.save()
+
+    return 201, prediction
 
 
 @router.put(
     "/predictions/{predict_id}",
     response={201: PredictionSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
     auth=django_auth,
+    tags=["registry", "predictions"],
+    include_in_schema=False,
 )
 def update_prediction(request, predict_id: int, payload: PredictionIn):
     try:
@@ -280,7 +339,11 @@ def update_prediction(request, predict_id: int, payload: PredictionIn):
         for attr, value in payload.dict().items():
             setattr(prediction, attr, value)
         # TODO: Add commit verification if commit has changed
-        prediction.save()
+
+        if not calling_via_swagger(request):
+            # Not realy required, since include_in_schema=False
+            prediction.save()
+
         return 201, prediction
     except Prediction.DoesNotExist:
         return 404, {"message": "Prediction not found"}
@@ -290,6 +353,8 @@ def update_prediction(request, predict_id: int, payload: PredictionIn):
     "/predictions/{predict_id}",
     response={204: SuccessSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
     auth=django_auth,
+    tags=["registry", "predictions"],
+    include_in_schema=False,
 )
 def delete_prediction(request, predict_id: int):
     try:
@@ -298,7 +363,10 @@ def delete_prediction(request, predict_id: int):
         if request.user != prediction.model.author.user:
             return 403, {"message": "You are not authorized to delete this prediction"}
 
-        prediction.delete()
+        if not calling_via_swagger(request):
+            # Not realy required, since include_in_schema=False
+            prediction.delete()
+
         return 204, {"message": f"Prediction {prediction.id} deleted successfully"}
     except Prediction.DoesNotExist:
         return 404, {"message": "Prediction not found"}
